+++ frontend/src/screens/CadastroScreen/index.tsx
@@ -5,16 +5,13 @@ import {
   TextInput,
   TouchableOpacity,
   Image,
-  StyleSheet,
   Alert,
 } from "react-native";
 import { StackNavigationProp } from "@react-navigation/stack";
 import { RootStackParamList } from "../../../types";
+import styles from "./styles";
 import { BACKEND_API_URL } from "@env";
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import axios from "axios";
-import moment from "moment";
-
 
 
 const logo = require("../../../assets/logo.png");
@@ -37,9 +34,10 @@ export default function CadastroScreen({ navigation }: Props) {
     try {
       const storedUser = await AsyncStorage.getItem("user")
       if (storedUser) {
-        navigation.navigate("Main", { user: JSON.parse(storedUser) });      }
+        navigation.navigate("Main", { user: JSON.parse(storedUser) });
+      }
     } catch (error) {
-      console.error("Erro ao obter dados do AsyncStorage:", error);
+      console.log("Erro ao obter dados do AsyncStorage:", error);
     }
   };
   useEffect(() => {
@@ -76,16 +74,16 @@ export default function CadastroScreen({ navigation }: Props) {
       });
 
       const data = await response.json(); // Captura a resposta em JSON
-      
+
       if (response.ok) {
         Alert.alert("Success", "Cadastro Realizado com suceeso.");
-        console.log("cadastro com sucesso",response)
+        console.log("cadastro com sucesso", response)
         navigation.navigate("Login");
       } else {
         // Caso a resposta não seja ok, exibe a mensagem de erro vinda do backend
         Alert.alert("Erro", data.message || "Erro ao cadastrar usuário. Tente novamente.");
       }
-    } catch (error:any) {
+    } catch (error: any) {
       console.log("Erro ao cadastrar usuário:", error.message);
       Alert.alert("Erro", "Erro ao cadastrar usuário. Tente novamente.");
     }
@@ -147,96 +145,4 @@ export default function CadastroScreen({ navigation }: Props) {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "white",
-  },
-  tituloContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginBottom: 0,
-    marginTop: 0,
-  },
-  tituloText: {
-    fontSize: 30,
-    fontFamily: "Roboto_700Bold",
-    fontWeight: "bold",
-    color: "#91C788",
-    marginRight: 10,
-  },
-  logo: {
-    width: 60,
-    height: 60,
-    resizeMode: "contain",
-  },
-  formContainer: {
-    width: "85%",
-    padding: 20,
-    borderRadius: 8,
-    borderWidth: 2,
-    borderColor: "#BBDEB5",
-    backgroundColor: "#BBDEB5",
-    alignItems: "center",
-    marginTop: 0,
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: "bold",
-    marginBottom: 10,
-    color: "#64A759",
-    textAlign: "center",
-  },
-  inputContainer: {
-    width: "100%",
-    marginBottom: 15,
-  },
-  label: {
-    fontSize: 16,
-    fontWeight: "bold",
-    color: "#64A759",
-    marginBottom: 5,
-  },
-  input: {
-    width: "100%",
-    padding: 10,
-    borderRadius: 5,
-    borderWidth: 1,
-    borderColor: "#fff",
-    backgroundColor: "white",
-  },
-  button: {
-    width: "100%",
-    paddingVertical: 15,
-    backgroundColor: "#91C788",
-    borderRadius: 5,
-    alignItems: "center",
-    marginTop: 20,
-  },
-  buttonText: {
-    color: "white",
-    fontSize: 18,
-    fontWeight: "bold",
-  },
-  rowContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginTop: 20,
-  },
-  login: {
-    color: "#91C788",
-    fontSize: 20,
-    fontWeight: "bold",
-    marginLeft: 3,
-    marginTop: 0,
-  },
-  login2: {
-    color: "#797878",
-    fontSize: 17,
-    marginRight: 5,
-    marginTop: 0,
-  },
-});
+//
