+++ frontend/types/index.ts
@@ -7,7 +7,7 @@ export interface RootStackParamList extends ParamListBase {
   Settings: undefined;
   Login: undefined;
   Cadastro: undefined;
-  Main: {user:IuserLogin};
+  Main: undefined;
   Onboarding: undefined;
   Agua: undefined;
   Alimentacao: undefined;
